callbot V2/requirements.txt
# Agno Framework
agno>=0.1.0
agno[openai]

# OpenAI
openai>=1.0.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Monitoring
prometheus-client>=0.19.0

# CLI & Display
colorama>=0.4.6

psycopg2-binary>=2.9.0